Validar acessos e informações

+++ src/data/usecases/category/deleteCategory-usecase.ts
@@ -1,6 +1,7 @@
 import { DeleteCategoryUseCaseInterface } from 'src/data/abstract/usecases/category/deleteCategoryUseCase-interface';
 import { Category } from 'src/domain/entities/category';
 import { CategoryRepositoryInterface } from 'src/infra/abstract/repositories/categoryRepositor-interface';
+import { InvalidParamError } from 'src/utils/errors/invalidParam-error';
 
 export class DeleteCategoryUseCase implements DeleteCategoryUseCaseInterface {
   private readonly repository: CategoryRepositoryInterface;
@@ -13,7 +14,13 @@ export class DeleteCategoryUseCase implements DeleteCategoryUseCaseInterface {
     categoryId: string,
     restaurantId: string,
   ): Promise<Category> {
-    const deleted = this.repository.delete(categoryId, restaurantId);
+    const fountEntity = await this.repository.getOne(categoryId, restaurantId);
+
+    if (fountEntity === null) {
+      throw new InvalidParamError('Id');
+    }
+
+    const deleted = await this.repository.delete(categoryId, restaurantId);
 
     return deleted;
   }
