+++ src/data/usecases/order/deleteOrder-usecase.ts
@@ -1,6 +1,7 @@
 import { DeleteOrderUseCaseInterface } from 'src/data/abstract/usecases/order/deleteOrderUseCase-interface';
 import { Order } from 'src/domain/entities/order';
 import { OrderRepositoryInterface } from 'src/infra/abstract/repositories/orderRepository-interface';
+import { InvalidParamError } from 'src/utils/errors/invalidParam-error';
 
 export class DeleteOrderUseCase implements DeleteOrderUseCaseInterface {
   private readonly repository: OrderRepositoryInterface;
@@ -9,8 +10,14 @@ export class DeleteOrderUseCase implements DeleteOrderUseCaseInterface {
     this.repository = repository;
   }
 
-  public execute(orderId: string, restaurantId: string): Promise<Order> {
-    const deleted = this.repository.delete(orderId, restaurantId);
+  public async execute(orderId: string, restaurantId: string): Promise<Order> {
+    const fountEntity = await this.repository.getOne(orderId, restaurantId);
+
+    if (fountEntity === null) {
+      throw new InvalidParamError('Id');
+    }
+
+    const deleted = await this.repository.delete(orderId, restaurantId);
 
     return deleted;
   }
