+++ src/data/usecases/product/deleteProduct-usecase.ts
@@ -1,6 +1,7 @@
 import { DeleteProductUseCaseInterface } from 'src/data/abstract/usecases/product/deleteProductUseCase-interface';
 import { Product } from 'src/domain/entities/product';
 import { ProductRepositoryInterface } from 'src/infra/abstract/repositories/productRepository-interface';
+import { InvalidParamError } from 'src/utils/errors/invalidParam-error';
 
 export class DeleteProductUseCase implements DeleteProductUseCaseInterface {
   private readonly repository: ProductRepositoryInterface;
@@ -9,8 +10,17 @@ export class DeleteProductUseCase implements DeleteProductUseCaseInterface {
     this.repository = repository;
   }
 
-  public execute(productId: string, restaurantId: string): Promise<Product> {
-    const deleted = this.repository.delete(productId, restaurantId);
+  public async execute(
+    productId: string,
+    restaurantId: string,
+  ): Promise<Product> {
+    const fountEntity = await this.repository.getOne(productId, restaurantId);
+
+    if (fountEntity === null) {
+      throw new InvalidParamError('Id');
+    }
+
+    const deleted = await this.repository.delete(productId, restaurantId);
 
     return deleted;
   }
