+++ src/data/usecases/review/deleteReview-usecase.ts
@@ -1,6 +1,7 @@
 import { DeleteReviewUseCaseInterface } from 'src/data/abstract/usecases/review/deleteReviewUseCase-interface';
 import { Review } from 'src/domain/entities/review';
 import { ReviewRepositoryInterface } from 'src/infra/abstract/repositories/reviewRepository-interface';
+import { InvalidParamError } from 'src/utils/errors/invalidParam-error';
 
 export class DeleteReviewUseCase implements DeleteReviewUseCaseInterface {
   private readonly repository: ReviewRepositoryInterface;
@@ -9,8 +10,17 @@ export class DeleteReviewUseCase implements DeleteReviewUseCaseInterface {
     this.repository = repository;
   }
 
-  public execute(reviewId: string, restaurantId: string): Promise<Review> {
-    const deleted = this.repository.delete(reviewId, restaurantId);
+  public async execute(
+    reviewId: string,
+    restaurantId: string,
+  ): Promise<Review> {
+    const fountEntity = await this.repository.getOne(reviewId, restaurantId);
+
+    if (fountEntity === null) {
+      throw new InvalidParamError('Id');
+    }
+
+    const deleted = await this.repository.delete(reviewId, restaurantId);
 
     return deleted;
   }
